Refactor PassThruTask test class. Make storeState method able to store message or remap specific data items from result.
